+++ changelog.md
@@ -1,4 +1,10 @@
 # Changelog
 
-**./app/validator/**
-* Library files unchanged.
+**./app/common-errors.js**
+* Compressed text writing code for functions:
+	* writeRemoteIoPropertyErrorString
+	* writeRemoteIoPropertySupportString
+	* writeRemoteIoPropertyConstructString
+	* writeConnectDevicePropertyErrorString
+	* writeSetDeviceOutputWrongError
+	* writeUnexpectedTokenErrorString